requirements.txt
sgfmill>=0.9.4
runpod
supabase>=2.0.0
httpx[http2]
requests
pytest-asyncio>=0.23
python-dotenv>=1.0.0